Restructured files, cleaned up
