+++ components/Results/Results.module.css
@@ -1,25 +1,27 @@
+.main {
+  padding: 0rem;
+  margin: 0rem;
+}
+
+.titleBar {
+  padding: 1rem 0rem 0.5rem;
+  margin: 0rem 0rem 0.5rem 0rem;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+
+  width: 100%;
+  border-bottom: 1px solid var(--border-color);
+}
+
 .title {
-    padding: 0.25rem 0;
-    margin: 0;
+  padding: 0rem;
+  margin: 0rem;
 
-    color: #2571ff;
-    font-size: x-large;
-    text-align: left;
-}
-
-.details {
-    padding: 0;
-    margin: 0;
-
-    color: #4d4d4d;
-    font-size: small;
-    text-align: left;
-}
-
-.desc {
-    padding: 0.5rem 0 0.25rem 0;
-    margin: 0;
-
-    font-size: medium;
-    text-align: left;
+  font-size: 1.5rem;
+  font-weight: bold;
+  text-align: center;
+  color: var(--text-color);
 }
