--- components/Search/Search.module.css
@@ -1,106 +1,0 @@
-.movielist {
-  padding: 0rem;
-  margin: 1rem 0rem 0rem 0rem;
-
-  border-style: hidden;
-
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  justify-content: flex-start;
-  align-items: stretch;
-
-  gap: 1rem;
-
-  font-size: 1rem;
-  font-weight: 500;
-  text-align: left;
-}
-
-.moviecard {
-  padding: 0.5rem;
-  margin: 0rem;
-
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  justify-content: flex-start;
-  align-items: flex-start;
-
-  text-align: center;
-  color: inherit;
-  text-decoration: none;
-
-  border: 2px solid var(--border-color);
-  border-radius: 10px;
-  background-color: var(--card-color);
-
-  transition: border-color 0.5s;
-}
-.moviecard:hover,
-.moviecard:focus {
-  border: 2px solid var(--highlight-color);
-  cursor: pointer;
-}
-
-.movieattributes {
-  padding: 0rem;
-  margin: 0rem;
-
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  flex-grow: 5;
-  flex-shrink: 3;
-  justify-content: flex-start;
-  align-items: flex-start;
-
-  text-align: center;
-  color: inherit;
-  text-decoration: none;
-}
-
-.moviedata {
-  padding: 0.1rem 1.5rem 0.1rem 1.5rem;
-
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  justify-content: flex-start;
-  align-items: flex-start;
-
-  text-align: left;
-}
-
-.title {
-  padding: 0.25rem 0;
-  margin: 0;
-
-  color: var(--accent-color);
-  font-size: x-large;
-  text-align: left;
-}
-
-.details {
-  padding: 0;
-  margin: 0;
-
-  color: var(--alt-text-color);
-  font-size: small;
-  text-align: left;
-}
-
-.desc {
-  padding: 0.5rem 0 0.25rem 0;
-  margin: 0;
-
-  color: var(--text-color);
-
-  font-size: medium;
-  text-align: left;
-}
-
-.poster {
-  padding: 0rem;
-  margin: 0rem;
-}
