+++ components/Searchbar/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./Searchbar";
+export { default as SearchButton } from "./SearchButton";
