--- components/Trending/Trending.module.css
@@ -1,92 +1,0 @@
-.main {
-  padding: 0rem;
-  margin: 0rem;
-}
-
-.titleBar {
-  padding: 1.5rem 0rem 0.5rem;
-  margin: 1rem 0rem 0.5rem 0rem;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-
-  width: 100%;
-  border-bottom: 1px solid var(--border-color);
-}
-
-.trending {
-  padding: 0rem;
-  margin: 0rem;
-
-  font-size: 1.5rem;
-  font-weight: bold;
-  text-align: center;
-  color: var(--text-color);
-}
-
-.movielist {
-  padding: 0rem;
-  margin: 0rem;
-
-  border-style: hidden;
-
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  align-items: stretch;
-
-  gap: 1rem;
-
-  font-size: 1rem;
-  font-weight: 500;
-  text-align: left;
-}
-
-.moviecard {
-  padding: 0.5rem 0rem;
-  margin: 0rem;
-
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  flex: 0 1 10%;
-  justify-content: flex-start;
-  align-items: center;
-
-  text-align: center;
-  color: var(--text-color);
-  text-decoration: none;
-
-  border: 2px solid var(--border-color);
-  border-radius: 10px;
-  background-color: var(--card-color);
-
-  transition: border-color 0.5s, color 0.5s;
-}
-.moviecard:hover,
-.moviecard:focus {
-  border: 2px solid var(--highlight-color);
-  color: var(--highlight-color);
-  cursor: pointer;
-}
-
-.movieattributes {
-  padding: 0rem;
-  margin: 0rem;
-
-  flex-grow: 5;
-  flex-shrink: 3;
-}
-
-.poster {
-  padding: 0rem;
-  margin: 0rem;
-}
-
-.moviedata {
-  padding: 0rem 0.5rem;
-  margin: 0rem;
-}
