+++ pages/search/[query].tsx
@@ -2,14 +2,14 @@
 import { GetServerSidePropsContext } from "next";
 
 // Functions
-import { MovieCardList } from "components/MovieCard";
+import Results from "components/Results";
 
 // API data
 const api_key = process.env.API_KEY;
 const api_url = process.env.API_URL_SEARCH;
 
 export default function Main({ data }) {
-  return <MovieCardList movies={data.results} variant={"grid"} />;
+  return <Results movies={data.results} title={"Search Results"} />;
 }
 
 export async function getServerSideProps(context: GetServerSidePropsContext) {
