+++ pages/search/index.tsx
@@ -1,5 +1,5 @@
 // Functions
-import Searchbar from "components/Search/Searchbar";
+import Searchbar from "components/Searchbar/Searchbar";
 
 export default function Main() {
   return <Searchbar />; // End return
